requirements.txt
torch
torchvision
transformers
scikit-learn
xgboost
pyyaml
tqdm
optuna
numpy
pandas